BUG: Quan es feia la multiplicació per trobar l'espai lliure, el casting estava mal fet i desbordava ( feia casting de la multiplicacio i no de cada factor a quint64 que es com s'havia de fer )

+++ starviewer/src/tools/harddiskinformation.cpp
@@ -64,7 +64,7 @@ quint64 HardDiskInformation::getTotalBytesPlataformEspecific(QString path)
 
     if ( statvfs(path.toAscii(), &fsd) == 0 )
     {
-        total = static_cast<quint64>( fsd.f_blocks * fsd.f_frsize );
+        total = static_cast<quint64>( fsd.f_blocks ) * static_cast<quint64>( fsd.f_frsize );
         existsError = false;
     }
 
@@ -99,7 +99,7 @@ quint64 HardDiskInformation::getFreeBytesPlataformEspecific(QString path)
 
     if ( statvfs(path.toAscii(), &fsd) == 0 )
     {
-        total = static_cast<quint64>( fsd.f_bavail * fsd.f_bsize );
+        total = static_cast<quint64>( fsd.f_bavail ) * static_cast<quint64>( fsd.f_bsize );
         existsError = false;
     }
